+++ src/Input/InputParamTrait.php
@@ -10,89 +10,94 @@ declare(strict_types=1);
 
 namespace Laminas\Cli\Input;
 
-use RuntimeException;
-use Symfony\Component\Console\Helper\HelperSet;
-use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Output\OutputInterface;
-
-use function sprintf;
-
 /**
- * Provide promptable input options to your command.
+ * Provide the majority of methods needed to implement InputParamInterface.
  *
- * Compose this trait in your symfony/console command in order to provide the
- * ability to define options that will result in interactive prompts when not
- * provided. Such parameters can be added using the `addParam()` construct. When
- * present, you can then use `$input->getParam($name)` to retrieve the value. If
- * the value was provided as an option, that value will be returned; otherwise,
- * it will prompt the user for the value.
+ * This trait provides definitions for all but the following methods of the
+ * InputParamInterface:
+ *
+ * - getOptionMode()
+ * - getQuestion()
+ *
+ * Additionally, it defines the `$name` property, allowing implementations to
+ * set it in their constructors without needing to define the property
+ * themselves.
  */
 trait InputParamTrait
 {
-    /**
-     * @var array<string, InputParamInterface>
-     */
-    private $inputParams = [];
+    /** @var mixed */
+    private $default;
+
+    /** @var string */
+    private $description = '';
 
     /**
-     * @param null|mixed $default
-     * @return $this
-     * @throws RuntimeException
+     * Parameter name; must be set by class composing trait!
+     *
+     * @var string
      */
-    final public function addParam(InputParamInterface $param): self
+    private $name;
+
+    /** @var bool */
+    private $required = false;
+
+    /** @var null|string */
+    private $shortcut = null;
+
+    /**
+     * Default value to use if none provided.
+     *
+     * @return null|mixed
+     */
+    public function getDefault()
     {
-        if (! is_array($this->inputParams)) {
-            throw new RuntimeException(sprintf(
-                'Command %s uses $inputParams property; please do not override that property when using %s,'
-                . ' as it becomes incompatible with input parameter usage',
-                static::class,
-                InputParamTrait::class
-            ));
-        }
+        return $this->default;
+    }
 
-        $name = $param->getName();
+    public function getDescription(): string
+    {
+        return $this->description;
+    }
 
-        $this->addOption(
-            $name,
-            $param->getShortcut(),
-            $param->getOptionMode(),
-            $param->getDescription()
-            // default null, on purpose
-        );
+    public function getName(): string
+    {
+        return $this->name;
+    }
 
-        $this->inputParams[$name] = $param;
+    public function getShortcut(): ?string
+    {
+        return $this->shortcut;
+    }
 
+    public function isRequired(): bool
+    {
+        return $this->required;
+    }
+
+    /**
+     * @param mixed $defaultValue
+     */
+    public function setDefault($defaultValue): self
+    {
+        $this->default = $defaultValue;
         return $this;
     }
 
-    final public function run(InputInterface $input, OutputInterface $output)
+    public function setDescription(string $description): self
     {
-        parent::run(
-            new ParamAwareInput(
-                $input,
-                $output,
-                $this->getHelperSet()->get('question'),
-                $this->inputParams
-            ),
-            $output
-        );
+        $this->description = $description;
+        return $this;
     }
 
-    /**
-     * @param string|array|null $shortcut
-     * @param null|mixed $default
-     * @return $this
-     */
-    abstract public function addOption(
-        string $name,
-        $shortcut = null,
-        ?int $mode = null,
-        string $description = '',
-        $default = null
-    );
+    public function setShortcut(?string $shortcut): self
+    {
+        $this->shortcut = $shortcut;
+        return $this;
+    }
 
-    /**
-     * @return HelperSet
-     */
-    abstract public function getHelperSet();
+    public function setRequiredFlag(bool $required): self
+    {
+        $this->required = $required;
+        return $this;
+    }
 }
